requirements.txt
paddlepaddle-gpu==1.1.0
paddlepaddle-gpuvision==0.2.0
tqdm==4.28.1
scikit-image==0.15.0
scikit-learn==0.21.3
PyYAML==5.1
pyzmq==18.1.0
regex==2019.6.8
Pillow==6.2.2
pip==20.1
opencv-python==3.1.0.0
opt-einsum==3.1.0
packaging==19.2
pandas==0.17.0
notebook==6.0.1
numpy==1.16.0
natsort==7.0.0
matplotlib==3.0.3
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab==1.2.0
jupyterlab-server==1.0.6
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.0.9
imageio==2.6.1
imagesize==1.2.0
tensorflow-gpu==1.15.2
tensorboard==1.15.0
six==1.12.0
scipy==1.2.1
